backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.48
asyncpg==0.30.0
alembic==1.14.1
redis==5.2.1
hiredis==3.1.1
celery==5.4.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
pywebpush==2.0.1
